third_party/WebKit/LayoutTests/media/track/track-webvtt-tc000-empty.html

<!DOCTYPE html>
<title>Tests that an empty file is not recognized as a WebVTT file.</title>
<video>
    <track src="captions-webvtt/tc000-empty.vtt" default>
</video>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
async_test(function(t) {
    var track = document.querySelector("track");

    track.onload = t.step_func_done(function() {
        assert_equals(track.track.cues.length, 0);
    });
});
</script>
